+++ src/utils/render.js
@@ -1,0 +1,18 @@
+import {RenderPosition} from "../const.js";
+
+export const createElement = (template) => {
+  const newElement = document.createElement(`div`);
+  newElement.innerHTML = template;
+  return newElement.firstChild;
+};
+
+export const render = (container, element, place = RenderPosition.BEFOREEDN) => {
+  switch (place) {
+    case RenderPosition.AFTERBEGIN:
+      container.prepend(element);
+      break;
+    case RenderPosition.BEFOREEDN:
+      container.append(element);
+      break;
+  }
+};
